requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
python-multipart
numpy